Updated _atoi() using type casting

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -23,7 +23,7 @@ int _atoi(char *s)
 	sign = _getsign(s, (unsigned int *)&i);
 	num = _getnum(s, (unsigned int *)&i);
 
-	return (num * sign);
+	return ((int) (num * sign));
 }
 
 /**
